+++ tango/io/vfs/FileFolder.d
@@ -363,7 +363,7 @@ class FileGroup : VfsFiles
 
         ***********************************************************************/
 
-        final uint files ()
+        final size_t files ()
         {
                 return group.length;
         }
@@ -468,7 +468,7 @@ private class FolderGroup : VfsFolders
 
         ***********************************************************************/
 
-        final uint folders ()
+        final size_t folders ()
         {
                 if (members.length is 1)
                     return members[0].stats.folders;
@@ -481,7 +481,7 @@ private class FolderGroup : VfsFolders
 
         ***********************************************************************/
 
-        final uint entries ()
+        final size_t entries ()
         {
                 return files + folders;
         }
